# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lumo - CConnect",
    "version": "1.0.0",
    "description": "Lumo CConnect REST API: Auth, Users, Posts"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "description": "roleKey takes precedence over accountType (must match if both provided). Students (roleKey=3) require degreeId, yearOfJoining (aka academicYear), and studentId; Staff (roleKey=2) may omit degreeId. departmentId is required for all. gender is optional. fullName also accepts legacy alias name. Note: the field is roleKey, not roleId. Usernames must be 3-30 characters long and may contain letters, numbers, dot, underscore, or hyphen (no spaces).",
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRegisterResponse"
                },
                "example": {
                  "id": 42,
                  "name": "Alice Doe",
                  "username": "alice.doe",
                  "email": "alice@example.edu",
                  "degreeId": 101,
                  "departmentId": 5,
                  "designationId": 301,
                  "empId": "EMP-2024-0001",
                  "gender": "female",
                  "yearOfJoining": 2023,
                  "studentId": "STU-2023-0001"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "missingFields": {
                    "value": {
                      "message": "All fields are required"
                    }
                  },
                  "otpTicketRequired": {
                    "value": {
                      "message": "otpTicket is required (10 digits)"
                    }
                  },
                  "ticketInvalid": {
                    "value": {
                      "message": "ticket_invalid"
                    }
                  },
                  "ticketExpired": {
                    "value": {
                      "message": "ticket_expired"
                    }
                  },
                  "ticketUsed": {
                    "value": {
                      "message": "ticket_used"
                    }
                  },
                  "conflict": {
                    "value": {
                      "message": "roleKey and accountType conflict"
                    }
                  },
                  "unsupportedRole": {
                    "value": {
                      "message": "Unsupported roleKey"
                    }
                  },
                  "degreeRequired": {
                    "value": {
                      "message": "degreeId is required"
                    }
                  },
                  "invalidDegree": {
                    "value": {
                      "message": "Invalid degreeId"
                    }
                  },
                  "departmentRequired": {
                    "value": {
                      "message": "departmentId is required"
                    }
                  },
                  "invalidDepartment": {
                    "value": {
                      "message": "Invalid departmentId"
                    }
                  },
                  "roleNotConfigured": {
                    "value": {
                      "message": "Role not configured"
                    }
                  },
                  "invalidUsername": {
                    "value": {
                      "message": "Username must be 3-30 characters and can include letters, numbers, dot, underscore, or hyphen."
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "emailConflict": {
                    "value": {
                      "message": "Email already in use"
                    }
                  },
                  "usernameConflict": {
                    "value": {
                      "message": "Username already in use"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Registration failed"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegisterRequest"
              },
              "examples": {
                "student": {
                  "summary": "Student registration (roleKey=3)",
                  "value": {
                    "fullName": "Alice Doe",
                    "username": "alice.doe",
                    "email": "alice@example.edu",
                    "password": "secret",
                    "otpTicket": "1234567890",
                    "roleKey": 3,
                    "degreeId": 101,
                    "departmentId": 5,
                    "gender": "female",
                    "yearOfJoining": 2023,
                    "studentId": "STU-2023-0001"
                  }
                },
                "staff": {
                  "summary": "Staff registration (roleKey=2)",
                  "value": {
                    "fullName": "Prof Bob",
                    "username": "prof_bob",
                    "email": "bob@example.edu",
                    "password": "secret",
                    "otpTicket": "1234567890",
                    "roleKey": 2,
                    "departmentId": 2,
                    "designationId": 301,
                    "empId": "EMP-2024-0001",
                    "gender": "male"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "summary": "Create a new account"
      }
    },
    "/auth/login": {
      "post": {
        "description": "Login with email/username and password",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponse"
                },
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "user": {
                    "id": 42,
                    "name": "Alice Doe",
                    "username": "alice",
                    "email": "alice@example.edu",
                    "degreeId": 101,
                    "departmentId": 5
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Invalid credentials"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "summary": "Login with email/username and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginRequest"
              }
            }
          }
        }
      }
    },
    "/auth/request-otp": {
      "post": {
        "description": "Request OTP to email",
        "responses": {
          "200": {
            "description": "OTP sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleError"
                },
                "example": {
                  "error": "Invalid email"
                }
              }
            }
          },
          "403": {
            "description": "Domain not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleError"
                },
                "example": {
                  "error": "Email domain is not allowed"
                }
              }
            }
          },
          "409": {
            "description": "Already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleError"
                },
                "example": {
                  "error": "Email already registered. Please login."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "type": "integer"
                },
                "example": 120
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleError"
                },
                "example": {
                  "error": "OTP recently sent. Try again in 120s"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleError"
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          },
          "502": {
            "description": "Failed to send email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleError"
                },
                "example": {
                  "error": "Failed to send OTP email"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "summary": "Request OTP to email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequest"
              },
              "example": {
                "email": "student@example.edu"
              }
            }
          }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "description": "Verify OTP code",
        "responses": {
          "200": {
            "description": "Verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpVerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleError"
                },
                "examples": {
                  "missing": {
                    "value": {
                      "error": "email and otp are required"
                    }
                  },
                  "invalid": {
                    "value": {
                      "error": "Invalid OTP"
                    }
                  },
                  "expired": {
                    "value": {
                      "error": "OTP expired"
                    }
                  },
                  "none": {
                    "value": {
                      "error": "No active OTP. Request a new one."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleError"
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "summary": "Verify OTP code",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyRequest"
              },
              "example": {
                "email": "student@example.edu",
                "otp": "1234"
              }
            }
          }
        }
      }
    },
    "/auth/reference-data": {
      "get": {
        "description": "Provide roleKey=3 (Student) to receive degrees + departments; roleKey=2 (Staff) to receive designations + departments.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceDataResponse"
                },
                "examples": {
                  "student": {
                    "summary": "Student reference data",
                    "value": {
                      "roleKey": 3,
                      "degrees": [
                        {
                          "degreeId": 101,
                          "degreeAbbr": "BSc",
                          "degreeName": "Bachelor of Science",
                          "level": 1
                        }
                      ],
                      "departments": [
                        {
                          "departmentId": 5,
                          "departmentName": "Computer Science"
                        }
                      ]
                    }
                  },
                  "staff": {
                    "summary": "Staff reference data",
                    "value": {
                      "roleKey": 2,
                      "departments": [
                        {
                          "departmentId": 5,
                          "departmentName": "Computer Science"
                        }
                      ],
                      "designations": [
                        {
                          "designationId": 1,
                          "designationName": "Professor",
                          "isTeaching": true
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleError"
                },
                "example": {
                  "error": "Invalid or missing roleKey (2=Staff, 3=Student)"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleError"
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "roleKey",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "enum": [
                2,
                3
              ],
              "example": 3
            },
            "description": "2=Staff, 3=Student"
          }
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Fetch reference data by roleKey"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "message": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          },
          "example": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Bad request"
              }
            }
          }
        }
      },
      "SimpleError": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "error": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          },
          "example": {
            "type": "object",
            "properties": {
              "error": {
                "type": "string",
                "example": "Something went wrong"
              }
            }
          }
        }
      },
...
      "AuthRegisterRequest": {
        "type": "object",
        "description": "Payload for /auth/register. Usernames are trimmed and validated before duplicate checks.",
        "required": [
          "fullName",
          "username",
          "email",
          "password",
          "otpTicket"
        ],
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Display name shown in UI."
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "pattern": "^[A-Za-z0-9._-]+$",
            "description": "Case-sensitive handle. Trimmed before persistence. Allowed characters: letters, numbers, dot, underscore, hyphen.",
            "example": "alice.doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Trimmed and lowercased during registration."
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Minimum 8 characters recommended."
          },
          "otpTicket": {
            "type": "string",
            "pattern": "^[0-9]{10}$",
            "description": "10-digit ticket from /auth/verify-otp."
          },
          "roleKey": {
            "type": "integer",
            "enum": [
              2,
              3
            ],
            "description": "2 = Staff, 3 = Student. Must align with accountType if present."
          },
          "accountType": {
            "type": "string",
            "nullable": true,
            "description": "Legacy string alias for roleKey."
          },
          "degreeId": {
            "type": "integer",
            "nullable": true
          },
          "departmentId": {
            "type": "integer",
            "nullable": true
          },
          "designationId": {
            "type": "integer",
            "nullable": true
          },
          "empId": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true,
            "description": "male | female | other | prefer_not_to_say"
          },
          "yearOfJoining": {
            "type": "integer",
            "nullable": true
          },
          "academicYear": {
            "type": "integer",
            "nullable": true
          },
          "studentId": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "OtpRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "required": {
            "type": "array",
            "example": [
              "email"
            ],
            "items": {
              "type": "string"
            }
          },
          "properties": {
            "type": "object",
            "properties": {
              "email": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "email"
                  },
                  "description": {
                    "type": "string",
                    "example": "Email to receive the OTP. Normalized to lowercase."
                  }
                }
              }
            }
          },
          "example": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "student@example.edu"
              }
            }
          }
        }
      },
      "OtpRequestResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "message": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Status message suitable for UI display."
                  }
                }
              },
              "email": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "email"
                  }
                }
              },
              "domain": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Brand/organization domain inferred from email."
                  }
                }
              },
              "expiresInSeconds": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "description": {
                    "type": "string",
                    "example": "Seconds until the OTP expires."
                  }
                }
              },
              "otp": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "4-digit OTP echoed in non-production environments or when explicitly enabled."
                  },
                  "nullable": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          },
          "example": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "OTP sent to email"
              },
              "email": {
                "type": "string",
                "example": "student@example.edu"
              },
              "domain": {
                "type": "string",
                "example": "example.edu"
              },
              "expiresInSeconds": {
                "type": "number",
                "example": 900
              },
              "otp": {
                "type": "string",
                "example": "1234"
              }
            }
          }
        }
      },
      "OtpVerifyRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "required": {
            "type": "array",
            "example": [
              "email",
              "otp"
            ],
            "items": {
              "type": "string"
            }
          },
          "properties": {
            "type": "object",
            "properties": {
              "email": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "email"
                  }
                }
              },
              "otp": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "4-digit code from email."
                  }
                }
              }
            }
          },
          "example": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "student@example.edu"
              },
              "otp": {
                "type": "string",
                "example": "1234"
              }
            }
          }
        }
      },
...
      "AuthRegisterResponse": {
        "type": "object",
        "description": "Response issued after successful registration.",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "degreeId": {
            "type": "integer",
            "nullable": true
          },
          "departmentId": {
            "type": "integer",
            "nullable": true
          },
          "designationId": {
            "type": "integer",
            "nullable": true
          },
          "empId": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "yearOfJoining": {
            "type": "integer",
            "nullable": true
          },
          "studentId": {
            "type": "string",
            "nullable": true
          }
        },
        "example": {
          "id": 42,
          "name": "Alice Doe",
          "username": "alice.doe",
          "email": "alice@example.edu",
          "degreeId": 101,
          "departmentId": 5,
          "designationId": 301,
          "empId": "EMP-2024-0001",
          "gender": "female",
          "yearOfJoining": 2023,
          "studentId": "STU-2023-0001"
        }
      },
...
      "AuthLoginRequest": {
        "type": "object",
        "description": "Credentials for /auth/login. emailOrUsername is trimmed before lookup.",
        "required": [
          "emailOrUsername",
          "password"
        ],
        "properties": {
          "emailOrUsername": {
            "type": "string",
            "description": "Email (case-insensitive) or username (case-sensitive).",
            "example": "alice@example.edu"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Plain password sent over HTTPS.",
            "example": "P@ssw0rd123!"
          }
        }
      },
      "AuthLoginResponse": {
        "type": "object",
        "description": "JWT plus lightweight user profile returned after successful login.",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT signed with server secret."
          },
          "user": {
            "type": "object",
            "$ref": "#/components/schemas/UserSummary"
          }
        },
        "example": {
          "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
          "user": {
            "id": 42,
            "name": "Alice Doe",
            "username": "alice",
            "email": "alice@example.edu",
            "degreeId": 101,
            "departmentId": 5,
            "designationId": null
          }
        }
      },
...
    }
  },
...
}